refactor(book): Service 분리
- open api와 관련된 서비스 분리
